+++ cps/db.py
@@ -1079,7 +1079,7 @@ class CalibreDB:
 
     def get_browseable_cc_columns(self, config):
         return [col for col in self.get_cc_columns(config, filter_config_custom_read=True)
-                if col.normalized and col.datatype in ('text', 'enumeration', 'rating')]
+                if getattr(col, 'normalized', False) and col.datatype in ('text', 'enumeration', 'rating')]
 
     # read search results from calibre-database and return it (function is used for feed and simple search
     def get_search_results(self, term, config, offset=None, order=None, limit=None, *join):
