+++ cps/templates/list.html
@@ -34,7 +34,9 @@
         <div class="col-xs-10 col-sm-10 col-md-11"><a id="list_{{loop.index0}}" href="{% if custom_column is defined %}{{url_for('web.custom_property_books', column_id=custom_column.id, book_id=entry[0].id )}}{% elif entry.format %}{{url_for('web.books_list', data=data, sort_param='stored', book_id=entry.format )}}{% else %}{{url_for('web.books_list', data=data, sort_param='stored', book_id=entry[0].id )}}{% endif %}">
           {% if data == 'rating' or (custom_column is defined and custom_column.datatype == 'rating') %}
            <div class="rating">
-          {% for number in range((entry.name if data == 'rating' else entry[0].name)|float|int) %}
+          {% set star_value = (entry.name if data == 'rating' else entry[0].name)|float|int %}
+          {% if star_value > 0 %}
+          {% for number in range(star_value) %}
              <span class="glyphicon glyphicon-star good"></span>
              {% if loop.last and loop.index < 5 %}
                {% for numer in range(5 - loop.index) %}
@@ -42,6 +44,7 @@
               {% endfor %}
             {% endif %}
           {% endfor %}
+          {% endif %}
           </div>
           {% else %}
           {% if entry.format %}
